refactor FlowUseCase.kt and UseCase.kt to use `context-receivers` feature
