+++ bindingadapters/build.gradle.kts
@@ -22,6 +22,7 @@ android {
 
     kotlinOptions {
         jvmTarget = "11"
+        freeCompilerArgs += "-Xcontext-receivers"
     }
 
     dataBinding {
