+++ core/build.gradle.kts
@@ -32,6 +32,7 @@ android {
 
     kotlinOptions {
         jvmTarget = "11"
+        freeCompilerArgs += "-Xcontext-receivers"
     }
 }
