+++ example-hilt/build.gradle.kts
@@ -29,6 +29,7 @@ android {
 
     kotlinOptions {
         jvmTarget = "11"
+        freeCompilerArgs += "-Xcontext-receivers"
     }
 }
