+++ example-minimal/build.gradle.kts
@@ -26,6 +26,7 @@ android {
 
     kotlinOptions {
         jvmTarget = "11"
+        freeCompilerArgs += "-Xcontext-receivers"
     }
 }
