+++ example/build.gradle.kts
@@ -40,6 +40,7 @@ android {
 
     kotlinOptions {
         jvmTarget = "11"
+        freeCompilerArgs += "-Xcontext-receivers"
     }
 
     configurations.all {
